reorganize viewers
