+++ vizapp/viewers/viewer.py
@@ -1,0 +1,5 @@
+class Viewer:
+
+    def __init__(self, vizapp):
+
+        self._vizapp = vizapp
