+++ vizapp/viewers/viewer1d.py
@@ -1,0 +1,8 @@
+from .viewer import Viewer
+
+class Viewer1D(Viewer):
+
+    def __init__(self, *args, **kwargs):
+
+        super().__init__(*args, **kwargs)
+
